+++ examples/vite-vike/pages/+Head.jsx
@@ -1,0 +1,8 @@
+export function Head() {
+    return (
+        <>
+            <title>Vite + Vike</title>
+            <meta name="description" content="Vite + Vike" />
+        </>
+    );
+}