+++ examples/vite-vike/pages/+config.js
@@ -1,0 +1,9 @@
+import vikeReact from "vike-react/config";
+
+/**
+ * @type {import("vike/types").Config}
+ */
+export default {
+    extends: vikeReact,
+    meta: {},
+};