+++ examples/vite-vike/pages/test.jsx
@@ -1,1 +1,0 @@
-export const Page = () => <div>test</div>;