+++ examples/vite-vike/vite.config.js
@@ -1,17 +1,8 @@
 import react from "@vitejs/plugin-react";
-import ssr from "vike/plugin";
+import vike from "vike/plugin";
 
 import faviconsPlugin from "@anolilab/unplugin-favicons/vite";
+
 export default {
-    plugins: [
-        react(),
-        faviconsPlugin(),
-        ssr({
-            prerender: {
-                noExtraDir: true,
-                parallel: 1, // Can be `number` or `boolean`
-                partial: false,
-            },
-        }),
-    ],
+    plugins: [react(), vike(), faviconsPlugin()],
 };