+++ taze.config.js
@@ -2,11 +2,7 @@ import { defineConfig } from "taze";
 
 export default defineConfig({
     // ignore packages from bumping
-    exclude: [
-        "@types/node",
-        "eslint",
-        "eslint-plugin-vitest",
-    ],
+    exclude: ["@types/node", "eslint", "eslint-plugin-vitest"],
     ignorePaths: ["node_modules"],
     mode: "latest",
     recursive: true,